Probabalistic distributed tracing
To demonstrate and test distributed tracing
across multiple opbeans and agent implementations,
we now support specifying a set of services
to which the "/api" routes will be proxied
with some configurable probability. By default,
50% of the time opbeans-go will service these
routes itself, and the rest of the time will
randomly choose a service specified in
OPBEANS_SERVICES to which the requests will be
proxied.